Validaciones simples

+++ app/src/main/java/uv/tc/cowtrol/RegistroRanchoActivity.kt
@@ -25,7 +25,7 @@ class RegistroRanchoActivity : AppCompatActivity() {
 
         binding.btRegistrarRancho.setOnClickListener{
             if (validarCampos()){
-                val nuevoRancho = Rancho(binding.etNombreRegistroRancho.text.toString(),binding.etUbicacion.text.toString(),binding.etTelefono.text.toString().toInt(),
+                val nuevoRancho = Rancho(binding.etNombreRegistroRancho.text.toString(),binding.etUbicacion.text.toString(),binding.etTelefono.text.toString(),
                     binding.etProvincia.text.toString())
                 agregarRancho(nuevoRancho)
             }
